Auto deploy HTML spec (#65)
Fix #66 

* Auto deploy HTML spec

* Comment out some things to test deploy

For testing, always assume we're going to deploy and do everything
except pushing the actual change.

* There are no images that need to be copied.

* Use correct encoded deploy key

The deploy_key.enc was somehow empty.  Encode it again.

* Let compile.sh copy everything to the out directory

Make compile.sh do the necessary copying to the out directory so that
deploy.sh doesn't have to worry about it.  compile.sh is responsible
for getting everything set up for the HTML version of spec.

* Revert previous changes meant for debugging

* Set commit author email to a real address